abstract out getEditModeKeyboardShortcuts getNormalModeKeyboardShortcuts setupKeyboardHandling
