+++ src/notebook/notebook.js
@@ -4,13 +4,7 @@ import { deserialize } from './deserialize';
 import { createCell } from './createCell';
 import { checkConsistency } from './checkConsistency';
 import { selectCell } from './selectCell';
-import { switchToCommandMode } from './switchToCommandMode';
-import { switchToEditMode } from './switchToEditMode';
-import { convertCellToCode } from './convertCellToCode';
-import { convertCellToHeading } from './convertCellToHeading';
-import { convertCellToMarkdown } from './convertCellToMarkdown';
-import { convertCellToRaw } from './convertCellToRaw';
-import { convertCellToScala } from './convertCellToScala';
+
 import { copyCell } from './copyCell';
 import { cutCell } from './cutCell';
 import { deleteCell } from './deleteCell';
@@ -24,7 +18,6 @@ import { insertNewCellBelow } from './insertNewCellBelow';
 import { insertNewScalaCellAbove } from './insertNewScalaCellAbove';
 import { insertNewScalaCellBelow } from './insertNewScalaCellBelow';
 import { appendCellAndRun } from './appendCellAndRun';
-import { moveCellDown } from './moveCellDown';
 import { moveCellUp } from './moveCellUp';
 import { mergeCellBelow } from './mergeCellBelow';
 import { splitCell } from './splitCell';
@@ -32,8 +25,6 @@ import { pasteCellAbove } from './pasteCellAbove';
 import { pasteCellBelow } from './pasteCellBelow';
 import { undoLastDelete } from './undoLastDelete';
 import { runCell } from './runCell';
-import { runCellAndInsertBelow } from './runCellAndInsertBelow';
-import { selectNextCell } from './selectNextCell';
 import { runCellAndSelectBelow } from './runCellAndSelectBelow';
 import { saveNotebook } from './saveNotebook';
 import { loadNotebook } from './loadNotebook';
@@ -43,10 +34,9 @@ import { toggleOutput } from './toggleOutput';
 import { editName } from './editName';
 import { saveName } from './saveName';
 import { toggleSidebar } from './toggleSidebar';
+import { moveCellDown } from './moveCellDown';
 // figured out how to use `export default function` syntax here
 // hence no {} curly braces
-import selectPreviousCell from './selectPreviousCell';
-import displayKeyboardShortcuts from './displayKeyboardShortcuts';
 import executeCommand from './executeCommand';
 import createNotebook from './createNotebook';
 import openNotebook from './openNotebook';
@@ -60,6 +50,9 @@ import createMenuItem from './createMenuItem';
 import setupMenus from './setupMenus';
 import createTool from './createTool';
 import toKeyboardHelp from './toKeyboardHelp';
+import setupKeyboardHandling from './setupKeyboardHandling';
+import getNormalModeKeyboardShortcuts from './getNormalModeKeyboardShortcuts';
+import getEditModeKeyboardShortcuts from './getEditModeKeyboardShortcuts';
 
 import { getObjectExistsRequest } from '../h2oProxy/getObjectExistsRequest';
 
@@ -183,259 +176,12 @@ export function notebook() {
       [createTool('question-circle', 'Assist Me', executeCommand(_, 'assist'))],
     ];
 
-
-    // (From IPython Notebook keyboard shortcuts dialog)
-    //
-    // The IPython Notebook has two different keyboard input modes.
-    // Edit mode allows you to type code/text into a cell
-    // and is indicated by a green cell border.
-    // Command mode binds the keyboard to notebook level
-    // actions and is indicated by a grey cell border.
-    //
-    // Command Mode (press Esc to enable)
-    //
-    const normalModeKeyboardShortcuts = [
-      [
-        'enter',
-        'edit mode',
-        switchToEditMode,
-      ],
-      // [ 'shift+enter', 'run cell, select below', runCellAndSelectBelow ]
-      // [ 'ctrl+enter', 'run cell', runCell ]
-      // [ 'alt+enter', 'run cell, insert below', runCellAndInsertBelow ]
-      [
-        'y',
-        'to code',
-        convertCellToCode,
-      ],
-      [
-        'm',
-        'to markdown',
-        convertCellToMarkdown,
-      ],
-      [
-        'r',
-        'to raw',
-        convertCellToRaw,
-      ],
-      [
-        '1',
-        'to heading 1',
-        convertCellToHeading(_, 1),
-      ],
-      [
-        '2',
-        'to heading 2',
-        convertCellToHeading(_, 2),
-      ],
-      [
-        '3',
-        'to heading 3',
-        convertCellToHeading(_, 3),
-      ],
-      [
-        '4',
-        'to heading 4',
-        convertCellToHeading(_, 4),
-      ],
-      [
-        '5',
-        'to heading 5',
-        convertCellToHeading(_, 5),
-      ],
-      [
-        '6',
-        'to heading 6',
-        convertCellToHeading(_, 6),
-      ],
-      [
-        'up',
-        'select previous cell',
-        selectPreviousCell,
-      ],
-      [
-        'down',
-        'select next cell',
-        selectNextCell,
-      ],
-      [
-        'k',
-        'select previous cell',
-        selectPreviousCell,
-      ],
-      [
-        'j',
-        'select next cell',
-        selectNextCell,
-      ],
-      [
-        'ctrl+k',
-        'move cell up',
-        moveCellUp,
-      ],
-      [
-        'ctrl+j',
-        'move cell down',
-        moveCellDown,
-      ],
-      [
-        'a',
-        'insert cell above',
-        insertNewCellAbove,
-      ],
-      [
-        'b',
-        'insert cell below',
-        insertNewCellBelow,
-      ],
-      [
-        'x',
-        'cut cell',
-        cutCell,
-      ],
-      [
-        'c',
-        'copy cell',
-        copyCell,
-      ],
-      [
-        'shift+v',
-        'paste cell above',
-        pasteCellAbove,
-      ],
-      [
-        'v',
-        'paste cell below',
-        pasteCellBelow,
-      ],
-      [
-        'z',
-        'undo last delete',
-        undoLastDelete,
-      ],
-      [
-        'd d',
-        'delete cell (press twice)',
-        deleteCell,
-      ],
-      [
-        'shift+m',
-        'merge cell below',
-        mergeCellBelow,
-      ],
-      [
-        's',
-        'save notebook',
-        saveNotebook,
-      ],
-      // [ 'mod+s', 'save notebook', saveNotebook ]
-      // [ 'l', 'toggle line numbers' ]
-      [
-        'o',
-        'toggle output',
-        toggleOutput,
-      ],
-      // [ 'shift+o', 'toggle output scrolling' ]
-      [
-        'h',
-        'keyboard shortcuts',
-        displayKeyboardShortcuts,
-      ],
-      // [ 'i', 'interrupt kernel (press twice)' ]
-      // [ '0', 'restart kernel (press twice)' ]
-    ];
-
-    if (_.onSparklingWater) {
-      normalModeKeyboardShortcuts.push([
-        'q',
-        'to Scala',
-        convertCellToScala,
-      ]);
-    }
-
-    //
-    // Edit Mode (press Enter to enable)
-    //
-    const editModeKeyboardShortcuts = [
-      // Tab : code completion or indent
-      // Shift-Tab : tooltip
-      // Cmd-] : indent
-      // Cmd-[ : dedent
-      // Cmd-a : select all
-      // Cmd-z : undo
-      // Cmd-Shift-z : redo
-      // Cmd-y : redo
-      // Cmd-Up : go to cell start
-      // Cmd-Down : go to cell end
-      // Opt-Left : go one word left
-      // Opt-Right : go one word right
-      // Opt-Backspace : del word before
-      // Opt-Delete : del word after
-      [
-        'esc',
-        'command mode',
-        switchToCommandMode,
-      ],
-      [
-        'ctrl+m',
-        'command mode',
-        switchToCommandMode,
-      ],
-      [
-        'shift+enter',
-        'run cell, select below',
-        runCellAndSelectBelow,
-      ],
-      [
-        'ctrl+enter',
-        'run cell',
-        runCell,
-      ],
-      [
-        'alt+enter',
-        'run cell, insert below',
-        runCellAndInsertBelow,
-      ],
-      [
-        'ctrl+shift+-',
-        'split cell',
-        splitCell,
-      ],
-      [
-        'mod+s',
-        'save notebook',
-        saveNotebook,
-      ],
-    ];
+    const normalModeKeyboardShortcuts = getNormalModeKeyboardShortcuts(_);
+    const editModeKeyboardShortcuts = getEditModeKeyboardShortcuts();
     const normalModeKeyboardShortcutsHelp = lodash.map(normalModeKeyboardShortcuts, toKeyboardHelp);
     const editModeKeyboardShortcutsHelp = lodash.map(editModeKeyboardShortcuts, toKeyboardHelp);
-    const setupKeyboardHandling = mode => {
-      let caption;
-      let f;
-      let shortcut;
-      let _i;
-      let _j;
-      let _len;
-      let _len1;
-      let _ref;
-      let _ref1;
-      for (_i = 0, _len = normalModeKeyboardShortcuts.length; _i < _len; _i++) {
-        _ref = normalModeKeyboardShortcuts[_i];
-        shortcut = _ref[0];
-        caption = _ref[1];
-        f = _ref[2].bind(this, _);
-        Mousetrap.bind(shortcut, f);
-      }
-      for (_j = 0, _len1 = editModeKeyboardShortcuts.length; _j < _len1; _j++) {
-        _ref1 = editModeKeyboardShortcuts[_j];
-        shortcut = _ref1[0];
-        caption = _ref1[1];
-        f = _ref1[2].bind(this, _);
-        Mousetrap.bindGlobal(shortcut, f);
-      }
-    };
     const initialize = () => {
-      setupKeyboardHandling('normal');
+      setupKeyboardHandling(_, 'normal');
       setupMenus(_, menuCell);
       Flow.Dataflow.link(_.load, loadNotebook.bind(this, _));
       Flow.Dataflow.link(_.open, openNotebook.bind(this, _));
